fix: 댓글 api 오류 수정 댓글 api 요청 안되는 오류 수정
