+++ src/lib/apis/commentApis.js
@@ -10,7 +10,7 @@ export const getCommentList = async (post_id) => {
 
 //댓글 작성 API
 export const postComment = async (post_id, commentContent) => {
-	const reqPath = `/post/${post_id}/${commentContent}`;
+	const reqPath = `/post/${post_id}/comments`;
 
 	const res = await axiosPrivate.post(reqPath, commentContent);
 	return res;
@@ -18,18 +18,18 @@ export const postComment = async (post_id, commentContent) => {
 
 //댓글 삭제 API
 
-export const deleteComment = async (post_id, commentContent, comment_id) => {
-	const reqPath = `/post/${post_id}/${commentContent}/:${comment_id}`;
+export const deleteComment = async (post_id, comment_id) => {
+	const reqPath = `/post/${post_id}/comments/${comment_id}`;
 
-	const res = await axiosPrivate.post(reqPath, commentContent);
+	const res = await axiosPrivate.delete(reqPath);
 	return res;
 };
 
 //댓글 신고 API
 
-export const reportComment = async (post_id, commentContent, comment_id) => {
-	const reqPath = `/post/${post_id}/${commentContent}/:${comment_id}/report`;
+export const reportComment = async (post_id, comment_id) => {
+	const reqPath = `/post/${post_id}/comments/${comment_id}/report`;
 
-	const res = await axiosPrivate.post(reqPath, commentContent);
+	const res = await axiosPrivate.post(reqPath);
 	return res;
 };
